+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This will break the projects developed on precedent versions.
 * DEL: Delete or deprecate some functions
 
 ## Current
+* MODIFY: references of conjugation forms: more short references
 
 ## Version 0.9.0 (2018-04-04)
 * REFACTORING